+++ lib/authrocket.rb
@@ -5,7 +5,7 @@ require 'jwt'
   require "authrocket/api/#{f}"
 end
 
-%w(app_hook auth_provider credential event login_policy membership org realm session user user_token).each do |f|
+%w(app_hook auth_provider credential event login_policy membership notification org realm session user user_token).each do |f|
   require "authrocket/#{f}"
 end