+++ lib/authrocket/event.rb
@@ -3,11 +3,13 @@ module AuthRocket
     crud :all, :find
 
     belongs_to :app_hook
+    belongs_to :auth_provider
     belongs_to :login_policy
     belongs_to :membership
     belongs_to :org
     belongs_to :realm
     belongs_to :user
+    has_many :notifications
 
     attr :event_type, :ip
     attr_datetime :event_at
@@ -19,6 +21,21 @@ module AuthRocket
       new(parsed, creds)
     rescue RecordNotFound
       nil
+    end
+
+    def notifications
+      reload unless @attribs[:notifications]
+      unless @stuffed_event
+        @attribs[:notifications].each do |n|
+          n.send :load, data: {event: self, event_id: id}
+        end
+        @stuffed_event = true
+      end
+      @attribs[:notifications]
+    end
+
+    def find_notification(nid)
+      notifications.detect{|n| n.id == nid} || raise(RecordNotFound)
     end
 
   end